Add LocallyTestableCode type & 2D grid implementation
Provide a 'LocallyTestableCode' interface, of block codes with a family
of 'local' tests. These are an ad hoc means of querying a small, random
sample of symbols from a test vector to establish whether it is close to
some codeword, with high confidence.

Also add a class of general concatenated linear codes, along with the
special subclass, 'GridLinearCode', of 2D cartesian products of fixed
pairs of codes. The latter forms a locally testable code (in our
terminology using _relative_ query size), with a local test consisting
of picking a row & column of the 2D grid at random & checking that each
is a valid codeword.