Add LocallyTestableProof type & 2D grid implementation
Provide a sub-interface of locally testable _proofs_. These are just
locally testable codes which include some nonlinear constraint on the
original message, termed the 'witness'. That is, every valid codeword
decodes to a witness satisfying a certain constraint. Thus a local test,
by establishing closeness to some codeword, shows that such a witness
_exists_ (with high confidence). In this way, locally testable proofs
become special cases of "probabilistically checkable proofs" (PCPs).

Also add a 'GridProof' class implementing 'LocallyTestableProof'. This
crucial construction is analogous to the earlier 'GridLinearCode', but
adds an algebraic constraint uniformly to (the decoding of) every row &
column of the 2D grid. In this way the witness constraint becomes a
(largely arbitrary) algebraic (2D) grid constraint itself, allowing the
encoding of any machine checkable proof.

To support the above & ensure local testability, add the subclass
'TripleLayerConstraint' of 'GridConstraint' (for witnesses) & the sub-
interface 'MultiplicativeLinearCode' for the top-layer row and column
codes of the 2D grid. Crucially, the RS code is 'multiplicative'.