Add ZeroKnowledgeLocallyTestableProof type & 2D grid implementation
Provide a sub-interface of "zero knowledge" locally testable proofs.
These have a randomised 'encode(..)' method from witnesses to codewords
& have specialised local tests extending 'ZeroKnowledgeLocalTest'.

We say that a local test is "zero knowledge" if the evidence (i.e. test
vector samples) it produces reveal nothing about the original witness.
Technically, this is achieved by giving it a 'simulate(..)' method that
produces fake evidence with the same probability distribution as that of
its inherited 'query(..)' method, which samples a real codeword.

Also add a 'ZeroKnowledgeGridProof' class (extending 'GridProof') which
implements this interface. To achieve this, the input 2D grid witness
is padded with a fixed number of random row and column codewords, which
have the effect of 'masking' ('XORing') out the witness data from the
set of sampled rows & columns (just like a one-time pad). This requires
a clear way to prevent the samples "passing through" the witness, so we
impose that the (top layer) row & column codes are systematic.

To aid the above, add 'isSystematic()' & 'codistance()' methods to
'Code' & 'LinearCode' respectively. The 'codistance' (i.e. distance of
the dual code) is the minimum size of sampled codeword points which can
be linearly dependent, thus measuring the 'masking' ability of the code.

Finally, add linear interpolation methods to 'Code' (which are stubbed
out by default, but implemented for Reed-Solomon codes). These are
necessary to implement the simulator for 'ZeroKnowledgeGridProof', which
must construct a set of randomised, fake row & column samples agreeing
wherever they cross.