Generify local test evidence & narrow 'localTest*(..)' return types
Add an evidence type parameter to 'LocallyTestableCode.LocalTest', to
improve usability & type safety of concrete tests, enforcing in
particular the requirement that 'query(..)' & 'simulate(..)' return the
same type of evidence (in the case of zero knowledge tests).

To further ease usability, narrow the return types of 'localTest*(..)'
in all the locally testable code (2D grid) implementations.