Negate 'Evidence.isFailure' to 'Evidence.isValid' for consistency
Use 'true' to mean success instead of failure for local tests, for
greater clarity & consistency with the 'Constraint' interface. Thus we
concern ourselves with type I (false positive) errors, instead of type
II (false negative) errors as before, taking closeness to a codeword (&
thus proof existence) to be the _alternative_ hypothesis & remoteness
from any codeword to be the _null_ hypothesis (which must be disproved).

As part of this, refer to the local test 'confidence' in place of its
'power' & its "false positive" instead of "false negative" probability.