Add ShortenedLinearCode class to simplify & decouple local tests
Simplify probability calculations in 'GridLinearCode.SimpleLocalTest' by
passing in a grid code of _shortened_ rows & columns, in the case of
zero knowledge proofs. This avoids the need for row & column exclusion
counts in the constructors, helping to decouple the class (a little)
from 'ZeroKnowledgeGridProof'.

Obtain a shortened code by zeroing out a fixed number of symbols from
the _end_ of each codeword, when defining 'ShortenedLinearCode'. This is
consistent with the implicit shortening done elsewhere (e.g. in
'PuncturedPolynomialCode'), along with our convention of embedding the
message at the end of a codeword, for systematic codes.

As the random row & column samples of a zero knowledge grid local test
are taken from the top & left sides of the grid (to avoid the witness),
which happens naturally when the outer row & column codes are shortened,
we longer need to override 'SimpleLocalTest.query(..)'. This further
simplifies 'ZeroKnowledgeGridProof' & makes it easier to customise.