Add compound test for grid codes, sampling fixed #s of rows & columns
Provide 'GridLinearCode.CompoundLocalTest' to replace the (generic)
repeated simple local tests that are currently used as compound tests
for 'GridLinearCode', 'GridProof' & 'ZeroKnowledgeGridProof'. It is
inherently more efficient, as it samples a fixed number of _distinct_
rows & columns (uniformly at random from all k-sets on n rows/columns,
with separate quotas for each dimension).

NOTE: It is also more efficient asymptotically than repeating a (biased)
random row/column sample, with the grid size & sample count -> infinity,
since making a _random_ choice between a row or a column leads to a high
count variance & associated probability of undersampling one or the
other, regardless of the total sample size.

As part of the above, add a 'BinomialUtils' class to compute k-out-of-n
random (index) choices, as well as ratios of binomial coefficients for
probability calculations. The latter uses a Lanczos approximation to the
log-Gamma function, to compute it anywhere it's defined on the reals.