Allow 'LowDiscrepancyFakeRandom' to be used in more cases without bias
Add an optional 'fieldsPerBound' parameter (== 1 by default) to the test
RNG 'LowDiscrepancyFakeRandom'. For choices greater than 1, this makes
each 'nextInt(n)' query cycle through a set of fields with independent
sqrt-prime multipliers in a round-robin fashion, instead of using the
same field each time. This eliminates bias when multiple consecutive
queries are made to 'nextInt(n)' for the same bound n, e.g. in the case
of a compound local grid test with '[row|column]SampleCount' > 1.