Move AES round function from test class to production
Add initial class to implement the AES-* block cipher as an algebraic
function (over the 8-bit AES field), based on the test circuitry used in
'ArithmeticCircuitTest'. For now, only implement the main round circuit,
not the key scheduling function (or final round). Move the remaining AES
test logic to 'AesTest'.

As part of this, add a missing 'getElements()' default method to the
'FiniteField' interface, to stream all the field elements.

TODO: Add key scheduling to finish the AES-128/192/256 implementation.