Fix bad parity check terms in ArithmeticCircuit::asFunction result
Fix a serious bug in the calculation of the parity check terms when
converting an arithmetic circuit to an algebraic function, caused by
a failure to offset by the absolute gate positions when emitting terms
of the form 'X_n - X_m' for each wire in the circuit. To do this, build
an offset map from gates to integers when iterating through them all.

Also fix an off-by-one bug in 'addWire[s](...)', caused by incorrect use
of 'Preconditions.checkPositionIndex' instead of 'checkElementIndex'.

Add regression tests for both the above.