Switch to big-endian 32/64 bit words in SHA-2 circuits
Simplify input/output of raw message blocks and the final digest to/from
'rawSha2StepCircuit()' (so that it is in bit-for-bit agreement with the
SHA-2 functions), by reversing the bit order of the individual 32 (resp.
64) bit words composing each narrow (resp. wide) message/state/IV block.

Accordingly, convert Sha2 'sumFn()', 'sigmaDeltaFn()' & 'roundConstFn()'
BooleanFunction constructions from little- to big- endian. To make this
easier and reduce the final number of XOR gates, add an auxiliary block
of partial carry bits to the former.

This also simplifies the unit tests by avoiding the need to reverse the
order of the numbers in every hex string & convert to/from little-endian
bit representations using 'Lists::reverse'.