Reduce AND gate count in Sha2.rawSha2StepCircuit()
Avoid adding the saved state to the current state twice, by using a
vectorial Choice gate to select a duplicated sum, instead of using a
mask to conditionally zero out the swapped states and then adding them
onto the un-swapped states pairwise (as done in the Groestl step circuit
where, unlike here, it is essentially free since it uses XOR, not ADD).

This only requires n*24 AND gates (field multiplications) instead of the
n*32 needed before, where n is the word size in bits (that is 32 or 64).

Also speed up Sha2 'choiceFn()' & 'majorityFn()' a little, by supplying
an explicit base function instead of using 'Builder::simpleBaseFn'.