Add BigIntegerRing & fix EuclideanDomain::gcdExt
Canonicalise the result of 'EuclideanDomain::gcdExt' so that the GCD is
always 'non-negative' (in a suitably general sense) and the x-component
of 'gcdExt(n, m)' is the canonical modular inverse of 'n / gcd(n, m)'
(modulo 'm / gcd(n, m)', so that it matches results of 'divMod').

To this end, add EuclideanDomain methods 'abs', 'signum' & 'invSignum',
so that the latter two give rise to units and 'abs(x) := x / signum(x)'.
Then "x is non-negative" means "x == abs(x)", which agrees with the
usual definition for integers, though not for e.g. Gaussian integers.

Also add a BigInteger ring to support arithmetic modulo large primes and
provide optimised versions of 'gcd' & 'gcdExt' which use the
corresponding fast BigInteger 'gcd' & 'modInverse' methods. Ensure that
the overridden methods produce results identical to those of super.