Add LargePrimeField & ReducibleGeneratorException classes
Enhance error handling of 'QuotientField::reciprocalOrZero' in the case
that the quotient ideal (generator) turns out to be reducible, causing
the required modular inversion to fail. Capture the discovered factors
using a new 'ReducibleGeneratorException' class. Use this to implement
a crude version of Lenstra's elliptic curve factorisation algorithm, as
a new test in 'WeierstrassCurveTest'.

Also add a 'LargePrimeField' class of big integers modulo p, for some
large prime p, extending 'QuotientField<BigInteger>'. In this way, it
can be made to implement 'FiniteField', which enables random sampling
and use in algebraic constraints such as 'ReedSolomonCode'. This also
requires choosing a primitive base for exponents and logarithms, so to
this end, provide simple private methods to find the smallest primitive
element of the field, from all the prime cofactors of the totient p - 1.