Fix zero doubling bug in ProjectiveWeierstrassCurve
Fix a regression in the previous commit which is triggered when the
neutral element (0 : 1 : 0) is doubled. Like the addition formula, the
doubling formula isn't complete & requires special handling at infinity.

Also fix a bug in 'isCurvePoint(..)' where (0 : 0 : 0) is considered a
valid curve point, even though that's not valid projective coordinates.

Add unit test cases to exercise the above bugs and additionally ensure
that all branches in the addition & doubling logic are covered.