Fix EuclideanDomain::gcdExt result canonicalisation
Swap round the signum and inverse signum factors which the first three
and last two fields of 'GcdExtResult' are multiplied by respectively, to
make the GCD 'non-negative'. Instead, the first three fields (including
the GCD itself) should be multiplied by the inverse signum of the
original GCD, not the last two.

This bug was not picked up because +/-1 are the only units in the tested
Euclidean domain (namely the ring of integers). Add a test of 'gcdExt'
on the ring of polynomials over GF(5), which is the simplest nontrivial
Euclidean domain readily to hand which has more than two units, to
exercise the bug.