Update dependencies + plugins & remove unstable API calls
Upgrade Guava to the latest release (32.0.1-jre), which stabilises much
of the API marked @Beta that our code was using, such as the methods
'ImmutableList::toImmutableList' & 'builderWithExpectedSize'. This makes
it easy to remove all the unstable API usage in production code except
that of the Guava graph API (used internally by 'ArithmeticCircuit'),
which is marked @Beta in its entirety. To do this, add the method
'StreamUtils::zip' as an (unoptimised) drop-in replacement for Guava's
'Streams::zip', which is still unfortunately marked @Beta.

(Unfortunately, updating Guava reintroduces most of the spurious might-
be-null warnings (plus some new ones) whose suppressions were removed in
the earlier commit, so (re-)add them where necessary.)

Also upgrade AutoValue to the latest version, which fixes a codegen bug
necessitating a workaround in 'BooleanFunction' with redundant abstract
overrides, to prevent a broken build. (And update JUnit.)

Finally, update the 'surefire', 'resources' & 'compiler' Maven plugins
to prevent build warnings when using the latest mvn release (3.9.2), and
add the Maven 'source' plugin to include sources with the installed
artifacts.