Add cloning + missing equals/hashCode & validation to finite fields
Provide a 'ShallowCopyable' abstract class, together with a static
method 'tryClone', which attempts to shallow copy any given structure,
such as a group, ring or field, returning the original object if copying
fails. This will be used later to support _canonicalised mappings_ from
the structure with weak tables, that is, precomputed data associated
with a given object without having to provide extra instance fields to
hold it, such as tables of roots of unity or other expensive-to-compute
cached data. Cloning a key in such a weak table, then computing the
associated data from the _clone_ instead of the original, breaks any
strong reference back to the key, which would otherwise prevent it being
garbage collected and cause a leak. Therefore, it is useful to be able
to clone (shallow copy) structures used as keys, even if they have value
semantics, as any associated data such as a table of ring elements will
typically point back to it.

As part of this, add missing 'equals' & 'hashCode' methods to all the
classes implemening 'Field', as finite fields will later use such a
canonicalised mapping to support taking square roots. Take care to
define 'equals(Object)' in a way most amenable to shallow copying and
anonymous subclassing, using the following form:

  public boolean equals(Object o) {
      return this == o || o instanceof Foo &&
          (o.getClass().isAssignableFrom(getClass())
               ? field1.equals(((Foo) o).field1) &&
                 field2.equals(((Foo) o).field2) &&
                 ...
               : o.equals(this));
  }

Thus 'equals(o)' checks that 'o.getClass()' is sandwiched between the
(named) static & (possibly anonymous) runtime class of the instance
before comparing fields, and defers to the 'equals' method of the
argument with a flipped call otherwise (unless the types are definitely
incompatible). With some care, this should not lead to infinite loops.

Also add missing validation to 'Gf256', 'Gf65536' & 'QuotientField' to
check that the field matches when doing any operation on a given element
(or elements). Additionally, verify primitivity of the generator
polynomial of the quadratic field extension in Gf65536 over its base
field in Gf256, in the constructor. This was missed out, leading to some
of the unit tests using a broken Gf65536 instance. Fix the tests as
appropriate and also reduce the amount of println output.

TODO: Add missing equals/hashCode for all the remaining structures.