Add PlusMinus<E> interface & FiniteField::(sqrt|invSqrt)
Provide an unsigned wrapper, 'PlusMinus<E>', round any Abelian group
element E, with a default implementation 'WitnessedPlusMinus<E>'. This
is used for representing results of square roots and (in a later commit)
the x-component of an elliptic curve point (or y-component in the case
of twisted Edwards curves), from which the curve point may be determined
only up to a sign. For convenience, make the witness element @Nullable
to permit a missing value, 'PlusMinus.ofMissing(group)', analoguous to
NaN. This makes the interface easier to use, since roughly half the time
taking a square root (or decoding a curve point from a single
coordinate) gives a result which cannot be represented without taking a
quadratic field extension. Canonicalise the +/- witness in the case of
'LargePrimeField', so that it always has a least significant bit of zero
(in its canonical representation as a big integer modulo p, for p > 2).
This will be important later when encoding/decoding curve points.

Also add default methods 'sqrt(E)'+'invSqrt(E)' to the 'FiniteField<E>'
interface. For efficiency, these require the precomputation of a table
of all the 2^k'th roots of unity for the given field. Create a
canonicalised mapping to a helper object, 'PowerTwoRootsOfUnits<E>',
holding the cached roots, with the cache implemented as a static weak
table. Use the 'ShallowCopyable::tryClone' method added in the previous
commit to break the strong references from values to keys that may
result from this.

Add tests for 'sqrt' & 'invSqrt' on 'Gf256' & 'Gf65536', where the
square roots are just inverse Frobenius automorphisms and hence linear
over GF(2). Also provide tests for them on 'LargePrimeField', with a new
test class 'LargePrimeFieldTest'.