Represent PlusMinus<P> for curve points P as a 1D projective point
Add a class 'P1PointCoordinates' of coordinates (X : Y) in 1D projective
space, analogous to the existing class 'P2PointCoordinates'. Use it to
represent elliptic curve points up to a sign, with abstract base class
'PlusMinusPoint<P, E>' implementing 'PlusMinus<P>', for points in P =
'Projective(Weierstrass|TwistedEdwards)Curve.Point<E>' (as well as
points in any future projective curves extending 'AbstractP2PointCurve',
such as Montgomery curves). The motivation for providing such a custom
implementation of 'PlusMinus<P>', instead of representing the unsigned
curve point directly with a witness, is to later support the Montgomery
ladder for fast scalar multiplication of unsigned (Montgomery) curve
points, represented with a single affine coordinate (or pair of
projective coordinates). These can multiplied by scalars for use in
Diffie-Hellman key exchange, but not added since the sign is unknown.

Add a public method 'AbstractP2PointCurve::plusMinusPoint' to create
unsigned curve points directly from a pair of 1D projective coordinates
(X : Z) (or (Y : Z) for twisted Edwards curve points). Retrieving the
witness from an instance of 'PlusMinusPoint' lazily evaluates it,
decompressing the curve point to something in 2D projective space. The
formulae to do this both involve taking a square root, which was
implemented in the previous commit, and therefore fail if the base field
is not an instance of 'FiniteField'.

Add unit tests for the implementations of 'PlusMinusPoint', as well as
curve point compression & decompression, for both the projective
Weierstrass and twisted Edwards curves. Additionally, test the
decompression of random 1D points and subsequent multiplication by the
prime group order L, to generate all 8 torsion points in Ed25519 (which
form a cycle) in 'ProjectiveTwistedEdwardsCurveTest'.