Fix flawed equals pattern for ShallowCopyable subclasses
Use slightly modified pattern for the 'equals(Object)' method body, for
subclasses of 'ShallowCopyable', in place of the flawed one introduced
in the earlier commit, which can cause infinite loops more easily than
expected (due to its use of recursion with a flipped 'equals' call). The
problem occurs in cases when neither object's class is assignable from
the other's. The new pattern used is as follows:

  public boolean equals(Object o) {
      return this == o || o instanceof Foo &&
          (o.getClass().isAssignableFrom(getClass())
               ? field1.equals(((Foo) o).field1) &&
                 field2.equals(((Foo) o).field2) &&
                 ...
               : o.equals(shallowCopy()));
  }

Here we pass 'shallowCopy()' to the flipped 'equals' method, instead of
'this' as before. This prevents an infinite loop that would otherwise
occur if there are two or more (say anonymous) subclasses of 'Foo' which
do not override 'equals' or 'hashCode', when calling 'equals' on their
respective instances. By making a shallow copy of the object passed in
the flipped call, we ensure that it is an instance of 'Foo' and not a
subclass thereof, so that its class is assignable from 'o.getClass()',
which makes the recursion stop at this point.

For the modified pattern to successfully prevent infinite loops, each
subclass of 'ShallowCopyable' should override 'equals' & 'shallowCopy'
at the same time, so that the class of the object returned by the latter
always matches the RHS of the 'instanceof' expression in the former.