Add Permutation abstract class & Structure interface
Move the 'AbelianGroup.select' interface method to a new 'Structure'
superinterface and swap its method parameters round to the more natural
(list, index) instead of (index, list). Also provide a (presently
unused) 'Structure.conditionalSwap' method, for doing constant time
conditional swaps, which will later be used to implement the Montgomery
ladder for suitable curves, plus constant time permutations of lists of
elements implementing 'Structure' using sorting networks.

Also provide a 'Permutation' abstract class with a final 'Swap' subclass
for use by the above, implementing 'java.util.function.IntUnaryOperator'
and defining 'equals', 'hashCode' & 'toString', with the latter showing
the permutation in standard mathematical notation as a product of cycles
(such as "(1, 2, 3)" or "(1, 4)(2, 3)" or "id"). Each instance of
'Permutation' is taken to be a mapping from ints to ints, such that all
but a finite number of integers map to themselves.

Add unit tests for the above and add 'org.hamcrest:hamcrest-library:2.2'
& 'com.github.seregamorph:hamcrest-more-matchers:0.1' to the set of test
dependencies, to make the tests a bit easier to write using matchers.